@@ -1,11 +1,11 @@
-package cn.netbuffer.zuuldemo.controller;
+package cn.netbuffer.gatewaydemo.controller;
 
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.core.env.Environment;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import javax.annotation.Resource;
+import jakarta.annotation.Resource;
 
 @RestController
 @RequestMapping("/info")
@@ -26,5 +26,4 @@ public class InfoController {
     public String environment(String property) {
         return environment.getProperty(property);
     }
-
 }
