@@ -1,7 +1,8 @@
 package cn.netbuffer.springclouddemo.userserviceinvoker;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.cloud.client.SpringCloudApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.cloud.client.loadbalancer.LoadBalanced;
 import org.springframework.cloud.openfeign.EnableFeignClients;
 import org.springframework.context.annotation.Bean;
@@ -13,7 +14,8 @@ import org.springframework.web.client.RestTemplate;
  * @SpringCloudApplication已经集成@SpringBootApplication/@EnableDiscoveryClient/@EnableCircuitBreaker
  */
 @EnableFeignClients
-@SpringCloudApplication
+@EnableDiscoveryClient
+@SpringBootApplication
 public class UserServiceInvokerApplication {
 
     /**
