@@ -1,6 +1,5 @@
 package cn.netbuffer.springclouddemo.userserviceinvoker.client;
 
-import com.google.common.collect.ImmutableMap;
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.stereotype.Component;
 import org.springframework.web.bind.annotation.*;
@@ -14,13 +13,13 @@ import java.util.Map;
 @FeignClient(value = "user-service-provider", fallback = UserClient.UserClientFallback.class)
 public interface UserClient {
 
-    Map nullUser = ImmutableMap.of("id", -1);
+    Map<String, Object> nullUser = Map.of("id", -1);
 
     @GetMapping("/user/{id}")
     String getUser(@PathVariable("id") Long id, @RequestParam Integer s);
 
     @PostMapping("/user")
-    Map add(@RequestBody Map user);
+    Map<String, Object> add(@RequestBody Map<String, Object> user);
 
     @Component
     class UserClientFallback implements UserClient {
@@ -30,9 +29,8 @@ public interface UserClient {
         }
 
         @Override
-        public Map add(Map user) {
+        public Map<String, Object> add(Map<String, Object> user) {
             return nullUser;
         }
     }
-
 }
