@@ -1,9 +1,11 @@
 package cn.netbuffer.userserviceprovider2;
 
 import org.springframework.boot.SpringApplication;
-import org.springframework.cloud.client.SpringCloudApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 
-@SpringCloudApplication
+@EnableDiscoveryClient
+@SpringBootApplication
 public class UserServiceProvider2Application {
 
     public static void main(String[] args) {
